:recycle: :: [#128] Feature / 모든 Feature에 Interface 적용
