+++ Projects/Feature/DetailNoticeFeature/Interface/Factory/DetailNoticeFactory.swift
@@ -1,0 +1,6 @@
+import BaseFeatureInterface
+import UIKit
+
+public protocol DetailNoticeFactory {
+    func makeViewController(noticeID: Int) -> any RoutedViewControllable
+}
