+++ Projects/Feature/DetailNoticeFeature/Sources/Assembly/DetailNoticeAssembly.swift
@@ -1,3 +1,4 @@
+import DetailNoticeFeatureInterface
 import NoticeDomainInterface
 import Swinject
 import UserDomainInterface
