--- Projects/Feature/DetailNoticeFeature/Sources/Factory/DetailNoticeFactory.swift
@@ -1,6 +1,0 @@
-import BaseFeature
-import UIKit
-
-public protocol DetailNoticeFactory {
-    func makeViewController(noticeID: Int) -> any StoredViewControllable
-}
