+++ Projects/Feature/HomeFeature/Interface/Factory/HomeFactory.swift
@@ -1,0 +1,7 @@
+import BaseFeatureInterface
+import Moordinator
+import UIKit
+
+public protocol HomeFactory {
+    func makeMoordinator() -> Moordinator
+}
