+++ Projects/Feature/HomeFeature/Sources/Assembly/HomeAssembly.swift
@@ -1,4 +1,5 @@
 import ConfirmationDialogFeatureInterface
+import HomeFeatureInterface
 import InputDialogFeatureInterface
 import MassageDomainInterface
 import MealDomainInterface
