+++ Projects/Feature/MassageFeature/Sources/Assembly/MassageAssembly.swift
@@ -1,4 +1,5 @@
 import MassageDomainInterface
+import MassageFeatureInterface
 import Swinject
 import UserDomainInterface
