--- Projects/Feature/MassageFeature/Sources/Factory/MassageFactory.swift
@@ -1,5 +1,0 @@
-import Moordinator
-
-public protocol MassageFactory {
-    func makeMoordinator() -> Moordinator
-}
