+++ Projects/Feature/MusicFeature/Sources/Factory/MusicFactoryImpl.swift
@@ -1,5 +1,6 @@
 import Moordinator
 import MusicDomainInterface
+import MusicFeatureInterface
 import ProposeMusicFeature
 import UserDomainInterface
