+++ Projects/Feature/MyViolationListFeature/Sources/Assembly/MyViolationListAssembly.swift
@@ -1,3 +1,4 @@
+import MyViolationListFeatureInterface
 import Swinject
 import ViolationDomainInterface
