+++ Projects/Feature/MyViolationListFeature/Sources/Factory/MyViolationListFactoryImpl.swift
@@ -1,4 +1,5 @@
 import BaseFeature
+import BaseFeatureInterface
 import MyViolationListFeatureInterface
 import ViolationDomainInterface
