+++ Projects/Feature/NoticeFeature/Sources/Factory/NoticeFactoryImpl.swift
@@ -1,7 +1,8 @@
 import ConfirmationDialogFeature
-import DetailNoticeFeature
+import DetailNoticeFeatureInterface
 import Moordinator
 import NoticeDomainInterface
+import NoticeFeatureInterface
 import UserDomainInterface
 
 struct NoticeFactoryImpl: NoticeFactory {
