+++ Projects/Feature/ProposeMusicFeature/Interface/Factory/ProposeMusicFactory.swift
@@ -1,0 +1,6 @@
+import BaseFeatureInterface
+import UIKit
+
+public protocol ProposeMusicFactory {
+    func makeViewController() -> any RoutedViewControllable
+}
