--- Projects/Feature/ProposeMusicFeature/Sources/Factory/ProposeMusicFactory.swift
@@ -1,6 +1,0 @@
-import BaseFeature
-import UIKit
-
-public protocol ProposeMusicFactory {
-    func makeViewController() -> any StoredViewControllable
-}
