+++ Projects/Feature/RenewalPasswordFeature/Interface/Factory/RenewalPasswordFactory.swift
@@ -1,7 +1,6 @@
-import BaseFeature
 import Moordinator
 import UIKit
 
-public protocol SignupFactory {
+public protocol RenewalPasswordFactory {
     func makeViewController(signinHandler: @escaping () -> Void) -> UIViewController
 }
