+++ Projects/Feature/RenewalPasswordFeature/Sources/Assembly/RenewalPasswordAssembly.swift
@@ -1,4 +1,5 @@
 import AuthDomainInterface
+import RenewalPasswordFeatureInterface
 import Swinject
 
 public final class RenewalPasswordAssembly: Assembly {
