+++ Projects/Feature/SignupFeature/Sources/Assembly/SignupAssembly.swift
@@ -1,4 +1,5 @@
 import AuthDomainInterface
+import SignupFeatureInterface
 import Swinject
 
 public final class SignupAssembly: Assembly {
