+++ Projects/Feature/SignupFeature/Sources/Factory/SignupFactoryImpl.swift
@@ -2,6 +2,7 @@ import AuthDomainInterface
 import BaseFeature
 import DesignSystem
 import DWebKit
+import SignupFeatureInterface
 import UIKit
 
 struct SignupFactoryImpl: SignupFactory {
