+++ Projects/Feature/SplashFeature/Interface/Factory/SplashFactory.swift
@@ -1,0 +1,6 @@
+import BaseFeatureInterface
+import UIKit
+
+public protocol SplashFactory {
+    func makeViewController() -> any RoutedViewControllable
+}
