--- Projects/Feature/SplashFeature/Interface/Interface.swift
@@ -1,1 +1,0 @@
-// This is for Tuist
