--- Projects/Feature/SplashFeature/Sources/Factory/SplashFactory.swift
@@ -1,6 +1,0 @@
-import BaseFeature
-import UIKit
-
-public protocol SplashFactory {
-    func makeViewController() -> any StoredViewControllable
-}
